refactor(queries): split 1344-LOC queries.ts — extract shared core (types/mappers/helpers) to queries/core.ts
Behavior-preserving: queries.ts re-exports core via barrel so all 194
consumers + type imports stay unchanged. Domain query functions remain in
queries.ts (838 LOC); reusable core is now a 541-LOC module.
typecheck 0, full suite 1065/1065 (== baseline), lint 0.
